Search by exclude color, frontpage color search block

+++ lib/core/lhgallery/lhhtmlvconverter.php
@@ -150,51 +150,56 @@ class erLhcoreClassHTMLVConverter {
         
         $config = erConfigClassLhConfig::getInstance();
 
-        $wwwUser = erConfigClassLhConfig::getInstance()->conf->getSetting( 'site', 'default_www_user' );
-   		$wwwUserGroup = erConfigClassLhConfig::getInstance()->conf->getSetting( 'site', 'default_www_group' );
-   		 
-		rename($pathExtracted,$photoDir.'/'.$fileNamePhysic);
-    	chown($photoDir.'/'.$fileNamePhysic,$wwwUser);
-    	chgrp($photoDir.'/'.$fileNamePhysic,$wwwUserGroup);
-    	chmod($photoDir.'/'.$fileNamePhysic,$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
-    	 
-    	$image->filesize = filesize($photoDir.'/'.$fileNamePhysic);
-    	$image->total_filesize = $image->filesize;
-    	$image->filepath = $params['photo_dir_photo'];
-
-    	$tag = new erLhcoreClassOgg( $photoDir.'/'.$fileNamePhysic);
-               
-        $image->pwidth = $tag->Streams['theora']['width'];
-        $image->pheight = $tag->Streams['theora']['height'];
+        if ($config->conf->getSetting( 'site', 'file_storage_backend' ) == 'filesystem')
+        {
+            $wwwUser = erConfigClassLhConfig::getInstance()->conf->getSetting( 'site', 'default_www_user' );
+       		$wwwUserGroup = erConfigClassLhConfig::getInstance()->conf->getSetting( 'site', 'default_www_group' );
+       		 
+    		rename($pathExtracted,$photoDir.'/'.$fileNamePhysic);
+        	chown($photoDir.'/'.$fileNamePhysic,$wwwUser);
+        	chgrp($photoDir.'/'.$fileNamePhysic,$wwwUserGroup);
+        	chmod($photoDir.'/'.$fileNamePhysic,$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
+        	 
+        	$image->filesize = filesize($photoDir.'/'.$fileNamePhysic);
+        	$image->total_filesize = $image->filesize;
+        	$image->filepath = $params['photo_dir_photo'];
+    
+        	$tag = new erLhcoreClassOgg( $photoDir.'/'.$fileNamePhysic);
+                   
+            $image->pwidth = $tag->Streams['theora']['width'];
+            $image->pheight = $tag->Streams['theora']['height'];
+            
+            $image->media_type = erLhcoreClassModelGalleryImage::mediaTypeHTMLV;
+            
+            if ($tag->Streams['picturable']) {
+                $image->has_preview = 1;
+                
+                $parts = explode('.',$fileNamePhysic);
+                array_pop($parts);
+                
+                if (isset($tag->Streams['theora']['framecount'])) {                
+                   $tag->GetPicture(round($tag->Streams['theora']['framecount']/2),$photoDir.'/original_'.implode('.',$parts).'.jpg');
+                } else {
+                   $tag->GetPicture(1,$photoDir.'/original_'.implode('.',$parts).'.jpg');  
+                }
+                
+                erLhcoreClassImageConverter::getInstance()->converter->transform( 'thumbbig', $photoDir.'/original_'.implode('.',$parts).'.jpg', $photoDir.'/normal_'.implode('.',$parts).'.jpg' ); 
+                erLhcoreClassImageConverter::getInstance()->converter->transform( 'thumb', $photoDir.'/original_'.implode('.',$parts).'.jpg', $photoDir.'/thumb_'.implode('.',$parts).'.jpg' ); 
+    
+                chown($photoDir.'/normal_'.implode('.',$parts).'.jpg',$wwwUser);
+                chgrp($photoDir.'/normal_'.implode('.',$parts).'.jpg',$wwwUserGroup);       
+                chmod($photoDir.'/normal_'.implode('.',$parts).'.jpg',$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
+                
+                chown($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$wwwUser);
+                chgrp($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$wwwUserGroup);     	
+                chmod($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
+                          
+                unlink($photoDir.'/original_'.implode('.',$parts).'.jpg');    // Delete original screenshot
+            }        
+        } elseif ($config->conf->getSetting( 'site', 'file_storage_backend' ) == 'amazons3') { 
         
-        $image->media_type = erLhcoreClassModelGalleryImage::mediaTypeHTMLV;
-        
-        if ($tag->Streams['picturable']) {
-            $image->has_preview = 1;
             
-            $parts = explode('.',$fileNamePhysic);
-            array_pop($parts);
-            
-            if (isset($tag->Streams['theora']['framecount'])) {                
-               $tag->GetPicture(round($tag->Streams['theora']['framecount']/2),$photoDir.'/original_'.implode('.',$parts).'.jpg');
-            } else {
-               $tag->GetPicture(1,$photoDir.'/original_'.implode('.',$parts).'.jpg');  
-            }
-            
-            erLhcoreClassImageConverter::getInstance()->converter->transform( 'thumbbig', $photoDir.'/original_'.implode('.',$parts).'.jpg', $photoDir.'/normal_'.implode('.',$parts).'.jpg' ); 
-            erLhcoreClassImageConverter::getInstance()->converter->transform( 'thumb', $photoDir.'/original_'.implode('.',$parts).'.jpg', $photoDir.'/thumb_'.implode('.',$parts).'.jpg' ); 
-
-            chown($photoDir.'/normal_'.implode('.',$parts).'.jpg',$wwwUser);
-            chgrp($photoDir.'/normal_'.implode('.',$parts).'.jpg',$wwwUserGroup);       
-            chmod($photoDir.'/normal_'.implode('.',$parts).'.jpg',$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
-            
-            chown($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$wwwUser);
-            chgrp($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$wwwUserGroup);     	
-            chmod($photoDir.'/thumb_'.implode('.',$parts).'.jpg',$config->conf->getSetting( 'site', 'StorageFilePermissions' ));
-                      
-            unlink($photoDir.'/original_'.implode('.',$parts).'.jpg');    // Delete original screenshot
-        }        
-    	
+        }
     }
     
     public static function handleUploadBatch(& $image,$params = array())
